requirements.txt
imageio==2.37.0
matplotlib==3.10.1
monai==1.4.0
numpy==2.2.3
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.1.0
prettytable==3.15.1
PyYAML==6.0.2
scipy==1.15.2
SimpleITK==2.4.1
statsmodels==0.14.4
torch==2.4.1
torch_summary==1.4.5
torchmetrics==1.4.2
torchvision==0.19.1
tqdm==4.66.5
vtk==9.4.1
